+++ CHANGELOG.md
@@ -3,10 +3,13 @@
 * Schema generation:
     * Compute defaults specified in `default_factory` (\#75).
     * Transform single-element `"allOf": [{"$ref": X}]` arrays into single `"$ref": X` key-value pair (\#79).
+    * Support providing `description` through `Field`, with priority over docstrings (\#89).
 * CLI commands:
     * Add `--verbose` option to `fractal-manifest check` (\#80).
     * In `fractal-manifest check`, display all manifest mismatches rather than the first one only (\#88).
 * Dependencies:
+    * Drop support for `python<3.11` (\#92).
+    * Drop support for `pydantic<2.11.0` (\#92).
     * Support `docstring-parser` v0.17 (\#76).
 
 # 0.3.0